Checks that is given number trimorphic or not
The program takes input from user and returns true or false based on the
input, that the given number is trimorphic or not. Trimorphic means the
cube of that number should ends with that number itself.
In future I would like to print all the trimorphic numbers between a
certain range or first N trimorphic numbers.